requirements.txt
crewai
langchain_groq
litellm
coqui-tts
torchcodec
langchain_community
streamlit
redis
qdrant-client
langchain
langchain-huggingface
sentence-transformers
requests
langchain-qdrant
psycopg2-binary
pandas
